Sync up with upstream main repo

+++ samples/gui/wasm-runtime-wgl/src/platform/linux/iwasm_main.c
@@ -48,7 +48,7 @@ static int baudrate = B115200;
 extern void init_sensor_framework();
 extern void exit_sensor_framework();
 extern void exit_connection_framework();
-extern int aee_host_msg_callback(void *msg, uint16_t msg_len);
+extern int aee_host_msg_callback(void *msg, uint32_t msg_len);
 extern bool init_connection_framework();
 
 #ifndef CONNECTION_UART
